move DataAdapter to its own module, misc cleanup
