+++ bokeh/charts/area.py
@@ -27,7 +27,8 @@ except ImportError:
     raise
 
 from ._chartobject import Builder, create_and_build
-from ..models import ColumnDataSource, Range1d, DataRange1d
+from ..models import ColumnDataSource, DataRange1d, GlyphRenderer, Range1d
+from ..models import Patch
 
 #-----------------------------------------------------------------------------
 # Classes and functions
